Flip image post FipFrame
